+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Unreleased_
 
 Fixed:
 
+- `Correctly set alarms during blocking processing <../../pull/209>`_
 - `Fix infinite loop when setting record's own value from on_update callback <../../pull/202>`_
 
 4.7.0_ - 2026-01-14